+++ packages/svelte/src/index.js
@@ -1,2 +1,4 @@
-import Link from './Link.html';
-export { Link };
+import Link from "./Link.html";
+import curiStore from "./store.js";
+
+export { Link, curiStore };